Refactor test helper schema merging
